Add modal to the drag and drop error
