+++ src/App.jsx
@@ -18,28 +18,13 @@ import Modal from './components/Basic/Modal/Modal';
 import LoadScreen from './components/Layouts/Body/Load/LoadScreen';
 
 function App() {
-  const [modalIsOpen, setIsOpen] = useState(false);
 
-  const openModal = () => {
-    setIsOpen(true);
-  };
-
-  const closeModal = () => {
-    setIsOpen(false);
-  };
 
   return (
     <Router>
       <div className="container-fluid m-0 p-0 row">
         <div className="col-3 p-0">
           <Menu />
-          <Button text="open modal" onClick={openModal} />
-          <Modal
-            isOpen={modalIsOpen}
-            closeModal={closeModal}
-            message="Error al intentar subir el archivo"
-            showButtons
-          />
         </div>
         <div className="col-9 p-0 bg-light">
           <Switch>
